added stuff

+++ transformer/train_convert.py
@@ -62,7 +62,8 @@ args = parser.parse_args()
 
 
 (x_train, y_train), (x_test, y_test) = datasets.cifar100.load_data()
-
+x_train = x_train.astype('float32')
+x_test = x_test.astype('float32')
 
         # Compute the mean and the variance of the training data for normalization.
 resize.layers[0].adapt(x_train)
